refactor(discord): split message-handler.ts into message-router + work-dispatch
Extract routing and command dispatch from the 1285-line message-handler.ts monolith:

- message-router.ts  — all routing logic (thread routing, mention handling, resume,
  permission checks, rate limiting, dedup). Owns MentionSessionInfo and
  MessageHandlerContext types.
- work-dispatch.ts   — handleWorkIntake and sendTaskResult (work task lifecycle and
  Discord embed callbacks).
- message-handler.ts — thin re-export entry point; all external imports unchanged.

No behavior changes. All 505 discord tests pass.
